requirements.txt
matplotlib==3.6.2
numpy==1.23.4
Pillow==10.3.0
prettymaps==0.1.3
osmnx==1.0.1
Shapely<2.0